Geometry Editor CLI 1.2.1-WIP 2025.01.18
Fixed wheels decompiling
Improved the hashing system
Improved configuration file error handling
Added the option to compress/decompress models without decompiling
Sped up RefPack compression by a bit
Changed texture priority handling
Added optional checks upon compiling
Made possible to use multiple filters in stream dumping
Fixed instability with stream replacing
Fixed UV maps 1 and 2 being ignored upon compiling
Fixed tangents calculation being ignored for some world materials
Fixed slightly wrong PositionW on certain vertex formats
Added the ability to compile multiple source files in a folder to a
single BIN containing multiple geometries
Added metadata blocks
Fixed issues with position markers

+++ src/fr/ni240sx/ucgt/geometryFile/io/VertexData3.java
@@ -1,11 +1,11 @@
 package fr.ni240sx.ucgt.geometryFile.io;
 
 public class VertexData3{
-	public double x;
-	public double y;
-	public double z;
+	public float x;
+	public float y;
+	public float z;
 
-	public VertexData3(double x, double y, double z) {
+	public VertexData3(float x, float y, float z) {
 		super();
 		this.x = x;
 		this.y = y;
@@ -13,8 +13,8 @@ public class VertexData3{
 	}
 
 	public VertexData3(String string, String string2, String string3) {
-		x = Double.parseDouble(string);
-		y = Double.parseDouble(string2);
-		z = Double.parseDouble(string3);
+		x = Float.parseFloat(string);
+		y = Float.parseFloat(string2);
+		z = Float.parseFloat(string3);
 	}
 }
